Fixed again profile screen and another file
Fixed again profile screen and another file
