+++ lib/screens/notification/notification_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_capstone/screens/notification/widget/card_all_booking_notification_widget.dart';
+import 'package:flutter_capstone/screens/notification/widget/card_booking_notification_widget.dart';
 import 'package:flutter_capstone/style/text_style.dart';
-import 'package:flutter_capstone/screens/notification/card_all_booking_notification_widget.dart';
-import 'package:flutter_capstone/screens/notification/card_booking_notification_widget.dart';
 
 class NotificationScreen extends StatefulWidget {
   const NotificationScreen({super.key});
