+++ lib/screens/sign_up/sign_up_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_capstone/widgets/sign_up/form_sign_up_widget.dart';
-import 'package:flutter_capstone/widgets/sign_up/header_sign_up_widget.dart';
+import 'package:flutter_capstone/screens/login/widget/header_login_widget.dart';
+import 'package:flutter_capstone/screens/sign_up/widget/form_sign_up_widget.dart';
 
 class SignUpScreen extends StatelessWidget {
   const SignUpScreen({super.key});
